Refactoring to isolate filters
