+++ app/views/companies/_filters.html.erb
@@ -124,8 +124,8 @@
                 </a>
               </li>
               <li>
-                <a href="" data-subcategory="employees" data-type="number">
-                  <%= t('filters.employees') %> <i class="glyphicon glyphicon-chevron-right"></i>
+                <a href="" data-subcategory="employees_count" data-type="number">
+                  <%= t('filters.employees_count') %> <i class="glyphicon glyphicon-chevron-right"></i>
                 </a>
               </li>
               <li>
@@ -142,8 +142,8 @@
             <ul>
               <li><a href="" data-type="equal"><%= t('filters.operators.equal') %></a></li>
               <li><a href="" data-type="contains"><%= t('filters.operators.contains') %></a></li>
-              <li><a href="" data-type="greater_than"><%= t('filters.operators.greater') %></a></li>
-              <li><a href="" data-type="less_than"><%= t('filters.operators.less') %></a></li>
+              <li><a href="" data-type="greater_than"><%= t('filters.operators.greater_than') %></a></li>
+              <li><a href="" data-type="less_than"><%= t('filters.operators.less_than') %></a></li>
             </ul>
           </div>
         </div>
