+++ app/views/layouts/application.html.erb
@@ -27,7 +27,7 @@
 	<script type="text/javascript" src="/libs/Modernizr/modernizr-2.8.3.min.js"></script>
 	<script type="text/javascript" src="/libs/Swiper/swiper.min.js"></script>
 	<script type="text/javascript" src="/libs/select-bootstrap/bootstrap-select.min.js"></script>
-	<script type="text/javascript" src="/js/main.js"></script>
+	<script type="module" src="/js/main.js"></script>
 	<script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js"></script>
 
 	<script>
