+++ config/locales/pt-br.yml
@@ -393,7 +393,7 @@ pt-br:
     name: 'Nome'
     location: 'Localização'
     description: 'Descrição'
-    employees: 'Empregados'
+    employees_count: 'Empregados'
     status: 'Status'
     category: 'Categoria'
     funding_type: 'Tipo de investimento'
@@ -413,8 +413,8 @@ pt-br:
     operators:
       equal: 'Igual'
       contains: 'Contém'
-      greater: 'Maior que'
-      less: 'Menor que'
+      greater_than: 'Maior que'
+      less_than: 'Menor que'
   activerecord:
     errors:
       messages:
