+++ public/js/main.js
@@ -1,3 +1,5 @@
+import { loadAutoComplete } from './filters.js';
+
 $(".container-add-your-company").hide();
 
 $(".btn-prev-signup").click(function(event) {
@@ -31,8 +33,8 @@ $(".btn-signup").click(function(event) {
 
 /* Disable sigin popup when clicks outside */
 $(document).click(function(event) {
-  $target = $(event.target);
-  $popup = $('.popup-login');
+  var $target = $(event.target);
+  var $popup = $('.popup-login');
 
   if(!$target.parents('.popup-login').length && $popup.is(':visible')) {
     $popup.removeClass('open-popup');
@@ -49,8 +51,8 @@ $(".btn-involved").click(function(event) {
 
 /* Disable involved popup when clicks outside */
 $(document).click(function(event) {
-  $target = $(event.target);
-  $popup = $('.popup-involved');
+  var $target = $(event.target);
+  var $popup = $('.popup-involved');
 
   if(!$target.parents('.popup-involved').length && $popup.is(':visible')) {
     $popup.removeClass('opened');
@@ -62,7 +64,7 @@ $(".btn-login-responsive").click(function(event) {
 });
 
 var peop = $(".all-peoples .line-people").size();
-max = 2;
+var max = 2;
 $('.all-peoples .line-people:lt(' + max + ')').show();
 
 $('.btn-show-more').click(function() {
@@ -121,7 +123,6 @@ $(".last-level-filter ul li a").click(function(event) {
   $(".control-breadcrumbs .bread .last-category").append(itemBread);
 });
 
-
 $(".sub-category li a").click(function(event) {
   var item_filter = '<div class="item-filter"><ul><li class="primary-nivel-filter"><button>' + $(this).text() + '</button></li><li class="second-nivel-filter"><select class="selectpicker"><option>input option 01</option><option>input option 02</option><option>input option 03</option></select></li><li class="last-nivel-filter"><input type="text" placeholder="Value"></li></ul><div class="button btn-remove-filter"><img src="/img/img-close-filter.png" alt=""></div></div>';
   if ($(this).find("i").length == 0) {
@@ -133,192 +134,12 @@ $(".sub-category li a").click(function(event) {
   return false;
 });
 
-function hideLastLevelFilters() { $('.last-level-filter li a').hide(); }
-function showLastLevelFilters() { $('.last-level-filter li a').show(); }
-
-function renderLastLevelFilters(clickedElement) {
-  const type = clickedElement.data().type;
-
-  if (type === 'select') {
-    /* Shows only the equals option */
-    hideLastLevelFilters();
-    $('.last-level-filter li a[data-type="equal"]').show();
-  } else if (['text', 'text-completion'].includes(type)) {
-    /* Shows only equals & contains options */
-    hideLastLevelFilters();
-    $('.last-level-filter li a[data-type="equal"]').show();
-    $('.last-level-filter li a[data-type="contains"]').show();
-  } else if (['number'].includes(type)) {
-    /* Hide only contains option */
-    showLastLevelFilters();
-    $('.last-level-filter li a[data-type="contains"]').hide();
-  }
-}
-
-function renderFilterInputs(index, subcategory, subcategoryType, type, category) {
-  let inputs = '';
-
-  inputs += '<input type="hidden" name="filter[' + index + '][type]" value="' + subcategory + '" />';
-  inputs += '<input type="hidden" name="filter[' + index + '][operator]" value="' + type + '" />';
-
-  if (['text', 'number'].includes(subcategoryType)) {
-    /* Allow free texting */
-    inputs += '<input type="text" name="filter[' + index + '][value]" data-category="' + category + '" ' +
-                     'data-subcategory="' + subcategory + '_input" data-type="' + type + '" />';
-  } else {
-    /* Adds i18n to it */
-    inputs += '<input type="hidden" name="filter[' + index + '][value]" data-category="' + category + '" ' +
-                     'data-subcategory="' + subcategory + '_input" data-type="' + type + '" />';
-    inputs += '<input type="text" data-category="' + category + '" data-subcategory="' + subcategory + '" ' +
-                     'data-type="' + type + '"/>';
-  }
-
-  return inputs;
-}
-
-$(".last-sub-category li a").click(function(event) {
-  showLastLevelFilters();
-
-  if ($(this).find("i").length > 0) {
-    $(".filter-geral").css({
-      transform: 'translateX(-199px) translateY(0px)',
-      transition: '.3s all'
-    });
-    $(".controls .arrow-left").removeClass('disabled');
-  }
-
-  renderLastLevelFilters($(this));
-});
-
-$(document).ready(function() {
-  function autoComplete(input) {
-    var subcategory = input.data()['subcategory'];
-    var cache = {};
-
-    input.autocomplete({
-      minLength: 0,
-      source: function(request, response) {
-        if (request.term in cache) {
-          response(cache[request.term])
-        } else {
-          var origin = $('a[data-subcategory="' + subcategory + '"]');
-
-          if (origin) {
-            var data = origin.data() || {};
-
-            if(data['autocompleteData']) {
-              const autocompleteData = data['autocompleteData'];
-              const values = Object.values(autocompleteData);
-
-              const result = values.filter((el, i) => {
-                return el.toLowerCase().includes(request.term.toLowerCase());
-              });
-              const key = Object.keys(autocompleteData)
-                                .find(k => autocompleteData[k] === result[0]);
-
-              /* Add autocomplete value to hidden input */
-              $('input[data-subcategory="' + subcategory + '_input"]').val(key);
-              response(result);
-            } else if(data['autocompleteSource']) {
-              $.ajax({
-                dataType: "json",
-                url: data['autocompleteSource'],
-                data: {
-                  term: request.term
-                },
-                success: function(result) {
-                  cache[request.term] = result;
-
-                  $('input[data-subcategory="' + subcategory + '_input"]').val(result);
-                  response(result);
-                },
-                error: function(_, status, error) {
-                  console.log(status, error)
-                }
-              });
-            }
-          }
-        }
-      }
-    });
-  }
-
-  function autocompleteFocus(givenInput) {
-    const subcategory = givenInput ? '' : $(".last-sub-category ul li.active a").data().subcategory;
-    const input = givenInput ? givenInput : $('input[data-subcategory="' + subcategory + '"]');
-
-    input.focus(function() {
-      autoComplete(input);
-      input.keydown();
-    });
-
-    autoComplete(input);
-    input.keydown();
-  }
-
-  /* Autocomplete on input focus, instead of new filters */
-  $(".last-nivel-filter input").click(function() { autocompleteFocus($(this)); });
-
-  /* Adds the filter and autocomplete other filters */
-  $(".last-level-filter li a").click(function(event) {
-    const category = $(".sub-category ul li.active a").text().trim();
-    const subcategory = $(".last-sub-category ul li.active a").data().subcategory;
-    const subcategoryLabel = $(".last-sub-category ul li.active a").text().trim();
-    const subcategoryType = $(".last-sub-category ul li.active a").data().type;
-    const type = $(".last-level-filter ul li.active a").data().type;
-    const typeLabel = $(".last-level-filter ul li.active a").text().trim();
-    const index = $('.item-filter').length;
-
-    const inputs = renderFilterInputs(index, subcategory, subcategoryType, type, category);
-    const item_filter = '<div class="item-filter"><ul><li class="primary-nivel-filter-secondary"><button>' + subcategoryLabel + '</button></li><li class="second-nivel-filter"><button>' + typeLabel + '</button></li><li class="last-nivel-filter">'+inputs+'</li></ul><div class="button btn-remove-filter"><img src="/img/img-close-filter.png" alt=""></div></div>';
-
-    if ($(this).find("i").length == 0) {
-      $("#modalFilter").modal("hide");
-      $(".filter-result").show();
-      $(".filter-complete").append(item_filter);
-      $('.selectpicker').selectpicker();
-
-      autocompleteFocus();
-    }
-
-    return true;
-  });
-})
-
-$(document).on("click", ".btn-remove-filter", function() {
-  $(this).parents(".item-filter").remove();
-});
-
-$(".add-new-filter").click(function(event) {
-  $(".filter-geral").css({
-    transform: 'translateX(0px) translateY(0px)',
-    transition: '.3s all'
-  });
-  $(".bread li span").empty();
-  $(".control-breadcrumbs .controls .arrow-left").addClass('disabled');
-  $(".level-filter ul li").removeClass('active');
-  $(".main-category ul li:first-child").addClass('active');
-
-});
-
-$(".btn-add-filter-modal").click(function(event) {
-  $(".filter-geral").css({
-    transform: 'translateX(0px) translateY(0px)',
-    transition: '.3s all'
-  });
-  $(".bread li span").empty();
-  $(".control-breadcrumbs .controls .arrow-left").addClass('disabled');
-  $(".level-filter ul li").removeClass('active');
-  $(".main-category ul li:first-child").addClass('active');
-});
-
 $(".arrow-left").click(function(event) {
   $(".filter-geral").css({
     transform: 'translateX(0px) translateY(0px)',
     transition: '.3s all'
   });
 });
-
 
 $(".btn-operator").click(function(event) {
   $(this).parents(".tag-operator").find('.container-operator').fadeToggle();
@@ -396,4 +217,13 @@ $(function() {
     changeMonth: true,
     changeYear: true,
   });
+});
+
+
+$(document).ready(function() {
+  loadAutoComplete();
+})
+
+$(document).on("click", ".btn-remove-filter", function() {
+  $(this).parents(".item-filter").remove();
 });
